requirements.txt
fastapi
matplotlib
uvicorn
psycopg2-binary
pandas
requests
python-multipart
pytest
starlette
tqdm
pydantic
pydantic_core
httpx
Jinja2
pytz
beautifulsoup4
openai
python-dotenv
plotly
kaleido
pillow
